Add files via upload
Contains several sheets with an analysis of the India - Brunei trade. It is an exercise from a course on International Trade Practices and Policies. The data contains the highest product categorization of 6 digits for the years 2001 to 2020. The product dominance is highlighted and Quantity and Unit Value Indices have been computed.
Furthermore, Averages are calculated and YoY growth has been listed. All these exercises have been done for imports and exports together.